day11 practice git branch workflow

+++ day11_gitBranch/calculator.h
@@ -2,4 +2,4 @@
 int add(int a,int b);
 int sub(int a,int b);
 int multiply(int a,int b);
-int divide(int a,int b);
+double divide(int a,int b);
